requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
openai==1.14.0
supabase==2.4.0
pydantic==2.6.4
python-dotenv==1.0.1
